chore(Browser): moved reactors to correct package
